poc: Preliminary support for file-level links
This is a hacky attempt at getting file ID links working. It does so by
adding a "phantom" root note to each notebook if it finds a `:ID:
<UUIDv4>` in the Book preface. The root note is given zero content and a
zeroed-out NotePosition linked to the Book. Finally, the Note gets a
`id` NoteProperty with the value extracted from `:ID:`.

It probably doesn't refresh as it should at this point and there are
probably a ton of other issues with this approach. However, initial
tests seems to work with my Org Roam notes.

Fixes orgzly/orgzly-android#901